requirements.txt
fpdf
pillow
pytesseract
awscli
boto3
svglib
reportlab
pdfkit